requirements.txt
crewai>=0.80.0
crewai-tools>=0.17.0
langchain-groq>=0.1.0
pandas>=2.1.4
numpy>=1.26.2
python-dotenv>=1.0.0
openpyxl>=3.1.2
pdfplumber>=0.10.3
plotly>=5.18.0
streamlit>=1.29.0
